Crate twilight_command_parser

👎 Deprecated since 0.8.1:
use interactions via twilight-http or twilight-gateway

twilight-command-parser

This crate has been deprecated. Please use Discord interactions via the twilight-gateway or twilight-http crates.

twilight-command-parser is a command parser for the twilight-rs ecosystem.

Included is a mutable configuration that allows you to specify the command names and prefixes. The parser parses out commands matching an available command and prefix and provides the command arguments to you.

Examples

A simple parser for a bot with one prefix ("!") and two commands: "echo" and "ping":

use twilight_command_parser::{Command, CommandParserConfig, Parser};

let mut config = CommandParserConfig::new();

config.add_command("echo", false);
config.add_command("ping", false);

// Add the prefix `"!"`.
// (Use `CommandParserConfig::add_prefixes` to add multiple prefixes)
config.add_prefix("!");

let parser = Parser::new(config);

// Now pass a command to the parser
match parser.parse("!echo a message") {
    Some(Command { name: "echo", arguments, .. }) => {
        let content = arguments.as_str();

        println!("Got an echo request to send `{}`", content);
    },
    Some(Command { name: "ping", .. }) => {
        println!("Got a ping request");
    },
    // Ignore all other commands.
    Some(_) => {},
    None => println!("Message didn't match a prefix and command"),
}

Re-exports

pub use self::config::CommandParserConfig;

Deprecated

Modules

configDeprecated

Configuration for a Parser.

Structs

ArgumentsDeprecated

An iterator over command arguments.

CommandDeprecated

Indicator that a command was used.

ParserDeprecated

A struct to parse prefixes, commands, and arguments out of messages.