Module twilight_command_parser::config

👎 Deprecated since 0.8.1:
use interactions via twilight-http or twilight-gateway

Configuration for a Parser.

Provided are methods for adding commands and removing them, as well as adding prefixes and removing prefixes. You can also iterate over commands and prefixes.

Structs

CommandParserConfigDeprecated

Configuration for a Parser.

CommandsDeprecated

Iterator over the parser configuration’s immutably borrowed commands.

CommandsMutDeprecated

Iterator over the parser configuration’s mutably borrowed commands.

PrefixesDeprecated

Iterator over the parser configuration’s immutably borrowed prefixes.

PrefixesMutDeprecated

Iterator over the parser configuration’s mutably borrowed prefixes.